refactor: update import statements and improve Rebar class documentation
